fix(core): surface serialization errors and enforce nested carrier determinism (#255)
## Summary

This PR resolves three determinism-related Core issues:

- **#65** — `SerializeFn` now returns `Result<Vec<u8>, String>` so
serialization failures are propagated instead of silently persisting
empty blobs.
- **#75** — The `BriocheExtensionType` derive macro now auto-collects
nested carriers from all struct/enum fields, filtering out primitives
and unwrapping common collection types. Every remaining nested carrier
must implement `BriocheExtensionType` (enforced via const assertions).
- **#77** — All persisted / extension-serialized structs now use
fixed-width integers (`u64`/`u32` in place of `usize`/`isize`).
In-memory use sites cast explicitly at the point of use.

To support the new macro assertions, `PolicyDecision`,
`TransitionTrace`, `DecisionNode`, and their transitive carriers now
derive `BriocheExtensionType` and `Default` where required.

## Invariant Impact

- `SerializeFn` returning `Result` closes the silent-empty-blob path.
- Auto-collection of nested carriers strengthens
`I-Eco-OrderedCollections` by making non-deterministic carriers a
compile-time error.
- Fixed-width integer schemas in persisted state make cross-architecture
replay deterministic, supporting `I-Core-Pure`.
- Cross-book change; ADR added at
`docs/architecture/adr/p3-core-determinism-65-75-77.md`.

## Human Checks (not enforced by CI)

- [x] Mechanism or policy? Policy stays in plugins/traits, never Core.
- [ ] New state machines have `proptest` coverage. *(No new state
machines; macro/serialization changes only.)*
- [x] Cross-book changes link an ADR.
- [x] Documentation matches code behavior.

## Verification

- `cargo fmt --check` — clean (only nightly-feature hints).
- `cargo clippy --workspace --all-targets --all-features -- -D warnings`
— clean.
- `cargo deny check all` — clean.
- `cargo-brioche-lint-invariants check-refs --root crates/` — clean.
- `cargo-brioche-lint-invariants check-matrix` — clean.
- `RUSTDOCFLAGS="-D warnings" cargo doc --workspace --no-deps` — clean.
- `cargo test --workspace --all-targets` — all passed.
- All commits GPG-signed.

Closes: #65
Closes: #75
Closes: #77

<!-- codesmith:footer -->
---
<a
href="https://app.blacksmith.sh/brio-labs/codesmith/brioche/pr/255"><picture><source
media="(prefers-color-scheme: dark)"
srcset="https://pr-comments-assets.blacksmith.sh/codesmith/view-with-codesmith-dark-v2.svg"><source
media="(prefers-color-scheme: light)"
srcset="https://pr-comments-assets.blacksmith.sh/codesmith/view-with-codesmith-light-v2.svg"><img
alt="View with Codesmith"
src="https://pr-comments-assets.blacksmith.sh/codesmith/view-with-codesmith-dark-v2.svg"></picture></a>
<a
href="https://backend.blacksmith.sh/track/enable-autofix?expires=1785929457&installation_id=142406161&pr_number=255&repository=brio-labs%2Fbrioche&return_to=https%3A%2F%2Fgithub.com%2Fbrio-labs%2Fbrioche%2Fpull%2F255&signature=56d15036c03ffdb44c60fe9b96b59efb487db5209f1ab0198b4582ab2b29f926"><picture><source
media="(prefers-color-scheme: dark)"
srcset="https://pr-comments-assets.blacksmith.sh/codesmith/autofix-with-codesmith-dark.svg"><source
media="(prefers-color-scheme: light)"
srcset="https://pr-comments-assets.blacksmith.sh/codesmith/autofix-with-codesmith-light.svg"><img
alt="Autofix with Codesmith"
src="https://pr-comments-assets.blacksmith.sh/codesmith/autofix-with-codesmith-dark.svg"></picture></a>
<sup>Need help on this PR? Tag <code>/codesmith</code> with what you
need. Autofix is disabled.</sup>

<!-- codesmith:autofix:disabled -->
<!-- /codesmith:footer -->

+++ crates/kernel/brioche-governance-default/src/tool_pipeline.rs
@@ -35,7 +35,7 @@ use crate::Priority;
 )]
 pub struct ToolResultFormatterState {
     /// Maximum size of a JSON result in bytes (0 = no limit).
-    pub max_result_bytes: usize,
+    pub max_result_bytes: u64,
     /// Total number of formatted results.
     pub formatted_count: u64,
 }
@@ -55,7 +55,7 @@ impl Default for ToolResultFormatterState {
 ///
 /// Refs: I-Core-ActiveToolCall
 pub struct ToolResultFormatter {
-    max_result_bytes: usize,
+    max_result_bytes: u64,
 }
 
 impl ToolResultFormatter {
@@ -70,7 +70,7 @@ impl ToolResultFormatter {
 
     /// Creates an instance with a custom size limit.
     /// Refs: I-Gov-TraitAtomic
-    pub fn with_max_result_bytes(max_result_bytes: usize) -> Self {
+    pub fn with_max_result_bytes(max_result_bytes: u64) -> Self {
         Self { max_result_bytes }
     }
 }
@@ -105,8 +105,9 @@ impl BriochePlugin for ToolResultFormatter {
         for result in results {
             let content = tool_outcome_to_string(&result.outcome);
 
-            if self.max_result_bytes > 0 && content.len() > self.max_result_bytes {
-                let meta = TruncatedToolResult::from_content(&content, self.max_result_bytes);
+            if self.max_result_bytes > 0 && content.len() > self.max_result_bytes as usize {
+                let meta =
+                    TruncatedToolResult::from_content(&content, self.max_result_bytes as usize);
                 let json = meta.to_json().map_err(|e| PluginError::Soft {
                     plugin_name: "tool_result_formatter".into(),
                     message: format!("JSON serialization failed: {e}"),
